Added `focus tab` command - aka numbered tab switching
Example: `focus tab 2` will switch to the 2nd (3rd) child of the
'closest' tabbed or stacked parent container.

This is useful e.g.: if you want to get rid of browser tabs and just
use sway's built-in tabs the whole time. Currently, the only options for movement are left/right or up/down.

Example keyboard binding:

    bindsym ctrl+1 focus tab 0
    bindsym ctrl+2 focus tab 1
    bindsym ctrl+3 focus tab 2
    bindsym ctrl+4 focus tab 3
    bindsym ctrl+5 focus tab 4
    bindsym ctrl+6 focus tab 5
    bindsym ctrl+7 focus tab 6
    bindsym ctrl+8 focus tab 7
    bindsym ctrl+9 focus tab 8